% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overshiny-package.R
\docType{package}
\name{overshiny-package}
\alias{overshiny}
\alias{overshiny-package}
\title{overshiny: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nicholasdavies/overshiny}
  \item Report bugs at \url{https://github.com/nicholasdavies/overshiny/issues}
}

}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
\keyword{internal}
