% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{overlayBounds}
\alias{overlayBounds}
\title{Align overlays with a ggplot2 plot}
\usage{
overlayBounds(ov, plot, xlim = c(NA, NA), ylim = c(NA, NA), row = 1L, col = 1L)
}
\arguments{
\item{ov}{A \code{\link[=reactiveValues]{reactiveValues()}} object returned by \code{\link[=overlayServer]{overlayServer()}}.}

\item{plot}{A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object used for overlay alignment.}

\item{xlim, ylim}{Vectors defining the coordinate limits for overlays.
Use \code{NA} to inherit axis limits from the plot panel.}

\item{row, col}{Row and column of the facet panel (if applicable).}
}
\value{
The plot object, to be returned from the \code{\link[shiny:renderPlot]{shiny::renderPlot()}} block.
}
\description{
Sets the pixel and coordinate bounds of the overlay area based on a
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object. This ensures that overlays are positioned
correctly in both visual and coordinate space. Currently only \code{ggplot2}
plots are supported.
}
\details{
Call this function within \code{\link[shiny:renderPlot]{shiny::renderPlot()}}, before returning the plot
object.
}
\examples{
\dontrun{
output$my_plot <- renderPlot({
    plot <- ggplot(df, aes(x, y)) + geom_line()
    overlayBounds(ov, plot, xlim = c(0, 100), ylim = c(0, NA))
})
}

}
\seealso{
\code{\link[=overlayServer]{overlayServer()}}
}
