% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overshiny-package.R
\docType{package}
\name{overshiny}
\alias{overshiny-package}
\alias{overshiny}
\title{Interactive overlays on Shiny plots}
\description{
\code{overshiny} provides draggable and resizable rectangular elements that
overlay plots in Shiny apps. This may be useful in applications where users
need to define regions on the plot for further input or processing.
Currently, the overlays are only designed to move along the x axis of the
plot.
}
\details{
The package exports a setup helper (\code{\link[=useOverlay]{useOverlay()}}),
UI components (\code{\link[=overlayToken]{overlayToken()}}, \code{\link[=overlayPlotOutput]{overlayPlotOutput()}}),
a server-side controller (\code{\link[=overlayServer]{overlayServer()}}), and a function for
aligning overlays to a ggplot2 or base plot (\code{\link[=overlayBounds]{overlayBounds()}}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nicholasdavies/overshiny}
  \item Report bugs at \url{https://github.com/nicholasdavies/overshiny/issues}
}

}
\author{
\strong{Maintainer}: Nick Davies \email{nicholas.davies@lshtm.ac.uk}

}
